Close #252 - Rename extras.render.syntax.render => extras.render.syntax
