--- modules/extras-render/shared/src/main/scala/extras/render/syntax/package.scala
@@ -1,10 +1,0 @@
-package extras.render
-
-/** @author Kevin Lee
-  * @since 2022-10-15
-  */
-package object syntax {
-  object render extends RenderSyntax
-
-  object all extends RenderSyntax
-}
